pyproject.toml
[tool.poetry]
name = "clustering"
version = "0.1.0"
description = "Supervised & unsupervised image data clustering"
authors = ["Evgeniy Trubnikov <parmesanshees@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
matplotlib = "^3.9.2"
pandas = "^2.2.3"
pathlib = "^1.0.1"
spacy = "^3.8.2"
pillow = "^11.0.0"

[tool.poetry.scripts]
main = 'src.clustering.__init__:main'
train_data_gen = 'src.train_data_img.__init__:main'

[tool.poetry.group.dev.dependencies]
mypy = "^1.13.0"
scipy-stubs = "^1.14.1.4"
rich = "^13.9.4"

[[tool.mypy.overrides]]
module = "sklearn.*"
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
